undo changes

+++ register/src/main/java/org/mainacad/register/repositories/TeacherRepository.java
@@ -4,5 +4,5 @@ import org.mainacad.register.domain.Teacher;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface TeacherRepository extends CrudRepository<Teacher,Long> {
+public interface TeacherRepository extends CrudRepository<Teacher,Integer> {
 }
